# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Pre Evaluador",
        "description": "Lista de todos las Rest Api's para Pre Evaluador",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "gerson.quezada@manuela.org.pe"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/Users/ListUsers": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Lista de Usuarios",
                "operationId": "List_Users",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/Users/ListBancos": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Lista de Bancos Comunales",
                "operationId": "List_Bancos",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/Users/ListPromotor": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Lista de Promotores",
                "operationId": "List_Promotores",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": " Use /auth to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}